chore: rename ButtonWithIcon to IconButton
