+++ packages/legacy/core/App/components/buttons/SettingsMenu.tsx
@@ -6,14 +6,14 @@ import { useTranslation } from 'react-i18next'
 import { RootStackParams, Screens, Stacks } from '../../types/navigators'
 import { testIdWithKey } from '../../utils/testable'
 
-import ButtonWithIcon, { ButtonLocation } from './ButtonWithIcon'
+import IconButton, { ButtonLocation } from './IconButton'
 
 const SettingsMenu: React.FC = () => {
   const navigation = useNavigation<StackNavigationProp<RootStackParams>>()
   const { t } = useTranslation()
 
   return (
-    <ButtonWithIcon
+    <IconButton
       buttonLocation={ButtonLocation.Left}
       accessibilityLabel={t('Screens.Settings')}
       testID={testIdWithKey('Settings')}
